added and tested AverageAccelaration, PenUpPause, VelocityInversion

+++ GrapholizeApp/src/main/java/New/Characteristics/CharacteristicNormalizedJerk.java
@@ -4,13 +4,13 @@ import New.Model.Entities.Dot;
 
 import java.util.List;
 
-public class CharacteristicNormalizedJerk extends Characteristic<Integer> {
+public class CharacteristicNormalizedJerk extends Characteristic<Double> {
 
     public CharacteristicNormalizedJerk(String name) {
         super(name);
     }
 
     @Override
-    public Integer calculate(List<List<Dot>> dots) {throw new UnsupportedOperationException();}
+    public Double calculate(List<List<Dot>> dots) {throw new UnsupportedOperationException();}
 
 }
